@@ -8,29 +8,33 @@
   "dependencies": {
     "@aspnet/signalr": "^1.1.0",
     "@types/chart.js": "2.7.42",
-    "@types/react": "^16.7.22",
-    "@types/react-dom": "^16.0.11",
+    "@types/react": "^16.8.2",
+    "@types/react-dom": "^16.8.0",
     "@types/react-redux": "^7.0.1",
     "@types/redux": "^3.6.0",
-    "@types/styled-components": "^4.1.6",
+    "@types/styled-components": "^4.1.8",
     "awesome-typescript-loader": "^5.2.1",
     "chart.js": "^2.7.3",
     "css-loader": "^2.1.0",
     "mini-css-extract-plugin": "^0.5.0",
-    "react": "^16.7.0",
+    "react": "^16.8.1",
     "react-chartjs-2": "2.7.4",
-    "react-dom": "^16.7.0",
+    "react-dom": "^16.8.1",
     "react-redux": "^6.0.0",
     "redux": "^4.0.1",
     "style-loader": "^0.23.1",
     "styled-components": "^4.1.3",
     "tslint": "^5.12.1",
     "tslint-react": "^3.6.0",
-    "typescript": "^3.3.1",
-    "webpack": "^4.29.0",
-    "webpack-cli": "^3.2.1"
+    "typescript": "^3.3.3",
+    "webpack": "^4.29.3",
+    "webpack-cli": "^3.2.3"
   },
   "devDependencies": {
-    "@types/node": "^10.12.21"
+    "@types/node": "^10.12.23"
+  },
+  "resolutions": {
+    "@types/react": "^16.8.2",
+    "@types/react-dom": "^16.8.0"
   }
 }
