@@ -3,8 +3,8 @@ import * as React from "react";
 import Main from "../components/Main";
 import SignalRContainer from "./signalr";
 
-export default class MainContainer extends React.Component<{}, {}> {
-  componentDidMount() {
+const MainContainer: React.FunctionComponent<{}> = () => {
+  React.useEffect(() => {
     // Patch in a font and use it everywhere (after it's done loading).
     const link = document.createElement("link");
     link.href = "https://fonts.googleapis.com/css?family=Open+Sans:400,700";
@@ -13,14 +13,17 @@ export default class MainContainer extends React.Component<{}, {}> {
       document.body.style.fontFamily = `"Open Sans", sans-serif`;
     };
     document.head.appendChild(link);
-  }
+    return () => {
+      document.head.removeChild(link);
+    };
+  }, []);
 
-  render() {
-    return (
-      <>
-        <Main />
-        <SignalRContainer />
-      </>
-    );
-  }
-}
+  return (
+    <>
+      <Main />
+      <SignalRContainer />
+    </>
+  );
+};
+
+export default MainContainer;
